# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Nexus Hosting API
  version: 1.0.0
  description: |
    Production-grade federated static site hosting network.
    Every node exposes this API. Authentication uses session cookies (browser)
    or Bearer API tokens (CLI / CI). Inter-node requests use Ed25519 signatures.
  contact:
    name: The No Hands Company
    url: https://github.com/The-No-Hands-company/Nexus-Hosting
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api
    description: Local node API
...
paths:
...
  /nodes/{id}/generate-keys:
    post:
      operationId: generateNodeKeys
      tags: [nodes]
      summary: Generate or rotate Ed25519 key pair for a node
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: New key pair generated (private key stored, not returned)
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodeId:
                    type: integer
                  publicKey:
                    type: string
                  message:
                    type: string
...
  /federation/handshake:
    post:
      operationId: initiateHandshake
      tags: [federation]
      summary: Initiate a signed handshake with a remote node
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetNodeUrl]
              properties:
                targetNodeUrl:
                  type: string
                  example: "https://node2.example.com"
      responses:
        "200":
          description: Handshake result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HandshakeResult"
...
  /federation/sync:
    post:
      operationId: federationSync
      tags: [federation]
      summary: |
        Receive a site_sync notification from a peer — fetch the file manifest
        and replicate all files to local object storage.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [siteDomain, deploymentId]
              properties:
                siteDomain:
                  type: string
                deploymentId:
                  type: integer
                timestamp:
                  type: string
      responses:
        "200":
          description: Sync complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FederationSyncResult"
        "202":
          description: Queued (manifest temporarily unavailable)
  /federation/manifest/{siteDomain}:
    get:
      operationId: getFederationManifest
      tags: [federation]
      summary: |
        Get the signed file manifest for a site. Used by peers to replicate
        files. Returns presigned download URLs valid for 1 hour.
      parameters:
        - name: siteDomain
          in: path
          required: true
          schema:
            type: string
        - name: deploymentId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Signed file manifest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FederationManifest"
        "404":
          $ref: "#/components/responses/NotFound"
  /federation/peers:
    get:
      operationId: listFederationPeers
      tags: [federation]
      summary: List known federation peers (paginated)
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Paginated peer list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedNodes"
  /federation/events:
    get:
      operationId: listFederationEvents
      tags: [federation]
      summary: Federation event log (paginated, newest first)
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Paginated event log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedEvents"

  # ── Gossip ───────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: sid
      description: Browser session cookie set by /login flow
    bearerToken:
      type: http
      scheme: bearer
      description: API token created via POST /tokens (prefix fh_)
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
        code:
          type: string
        status:
          type: integer
    PaginationMeta:
      type: object
      required: [total, page, limit, totalPages, hasNextPage, hasPrevPage]
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        hasNextPage:
          type: boolean
        hasPrevPage:
          type: boolean
...
    Node:
      type: object
      required: [id, name, domain, status, region, operatorName, operatorEmail, storageCapacityGb, bandwidthCapacityGb, joinedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        domain:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, maintenance]
        region:
          type: string
        operatorName:
          type: string
        operatorEmail:
          type: string
        storageCapacityGb:
          type: number
        bandwidthCapacityGb:
          type: number
        uptimePercent:
          type: number
        siteCount:
          type: integer
        publicKey:
          type: ["string", "null"]
        isLocalNode:
          type: integer
        joinedAt:
          type: string
        lastSeenAt:
          type: ["string", "null"]
        verifiedAt:
          type: ["string", "null"]
    PaginatedNodes:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Node"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
    SiteDeployment:
      type: object
      required: [id, siteId, version, status, fileCount, totalSizeMb, deployedAt]
      properties:
        id:
          type: integer
        siteId:
          type: integer
        version:
          type: integer
        status:
          type: string
          enum: [pending, active, failed, rolled_back]
        deployedBy:
          type: ["string", "null"]
        fileCount:
          type: integer
        totalSizeMb:
          type: number
        deployedAt:
          type: string
        createdAt:
          type: string
...
    HandshakeResult:
      type: object
      properties:
        success:
          type: boolean
        targetUrl:
          type: string
        discoveryData:
          type: ["object", "null"]
        pingResult:
          type: ["object", "null"]
        error:
          type: ["string", "null"]
...
    FederationSyncResult:
      type: object
      properties:
        status:
          type: string
          enum: [synced, skipped, queued, failed]
        siteDomain:
          type: string
        filesSync:
          type: integer
        failedFiles:
          type: integer
        originDomain:
          type: string
    FederationManifest:
      type: object
      properties:
        site:
          type: object
        deployment:
          $ref: "#/components/schemas/SiteDeployment"
        files:
          type: array
          items:
            type: object
            properties:
              filePath:
                type: string
              contentType:
                type: string
              sizeBytes:
                type: integer
              downloadUrl:
                type: string
        signature:
          type: ["string", "null"]
        servedBy:
          type: string
        expiresAt:
          type: string
    PaginatedEvents:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            type: object
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...